requirements.txt
pandas
numpy 
keras
matplotlib
Pillow
opencv-python #cv2
seaborn
scikit-image
tensorflow
mlflow

